Move TestProposalStatus
